Fixed memmory leaked issue where async was declared globally. Added linz.api.navigation and linz.api.model.getPermission. Applied view permission to templates
