+++ lib/api/configs.js
@@ -1,4 +1,5 @@
-var linz = require('../../');
+var linz = require('../../'),
+    clone = require('clone');
 
 /**
  * Get linz config by reference or by cloning a copy
