+++ lib/api/formtools/grid.js
@@ -1,4 +1,4 @@
-var linz = require('../../../');
+var linz = require('../../../'),
     async = require('async'),
     clone = require('clone');
