+++ lib/api/index.js
@@ -3,3 +3,4 @@ module.exports.url = require('./url');
 module.exports.model = require('./model');
 module.exports.views = require('./views');
 module.exports.configs = require('./configs');
+module.exports.navigation = require('./navigation');
