+++ lib/formtools/plugins/plugins-helpers.js
@@ -275,6 +275,7 @@ var exportDefault = function (exports) {
                 label: exp.label,
                 action: exp.action || 'export',
                 enabled: !(exp.enable === true),
+                permission: exp.permission || undefined,
                 exclusions: exp.exclusions || '_id,dateCreated,dateModified,createdBy,modifiedBy'
             };
