+++ lib/versions/renderers/cell/document-array.js
@@ -1,4 +1,5 @@
-var cellRenderers = require('./');
+var cellRenderers = require('./'),
+    async = require('async');
 
 module.exports = function documentArrayRenderer (val, record, fieldName, model, callback) {
