+++ views/modelIndex/header.jade
@@ -2,17 +2,18 @@
 	h1
 		a(href='' + linz.api.url.getAdminLink(model))= model.linz.formtools.model.plural
 .model-actions
-	if model.grid.canCreate && permissions.create === true
+	if model.grid.canCreate && permissions.create
 		.actions
 			span.create
 				a.btn.btn-primary(href='' + linz.api.url.getAdminLink(model, 'new'))
 					i.fa.fa-plus
 					| Create a new #{label.singular}
-	if model.grid.actions && model.grid.actions.length
+	if model.grid.actions && model.grid.actions.length && permissions.gridActions
 		.custom-actions
 			if model.grid.actions.length === 1
 				for action in model.grid.actions
-					a.btn.btn-default(href='' + linz.api.url.getAdminLink(model, action.action))= action.label
+					if action.permission && permissions[action.permission] || !action.permission
+						a.btn.btn-default(href='' + linz.api.url.getAdminLink(model, action.action))= action.label
 			else
 				.dropdown
 					button.btn.btn-default.dropdown-toggle(type='button', data-toggle='dropdown')
@@ -29,10 +30,12 @@
 									li(role='presentation')
 										a(role='menuitem', tabindex='-1', href='' + linz.api.url.getAdminLink(model, action.action))= action.label
 
-	if model.grid.groupActions && model.grid.groupActions.length
+	if model.grid.groupActions && model.grid.groupActions.length && permissions.groupActions
 		.group-actions.hidden
 			if model.grid.groupActions.length === 1
-				a.btn.btn-default(href='' + linz.api.url.getAdminLink(model, groupAction.action) data-linz-control=((groupAction.useModal) ? 'group-action' : false) data-target=((groupAction.useModal) ? '#groupActionModal' : false) )= groupAction.label
+				for groupAction in model.grid.groupActions
+					if groupAction.permission && permissions[groupAction.permission] || !groupAction.permission
+						a.btn.btn-default(href='' + linz.api.url.getAdminLink(model, groupAction.action) data-linz-control=((groupAction.useModal) ? 'group-action' : false) data-target=((groupAction.useModal) ? '#groupActionModal' : false) )= groupAction.label
 			else
 				.dropdown
 					button.btn.btn-default.dropdown-toggle(type='button', data-toggle='dropdown')
@@ -54,11 +57,11 @@
 				.toolbarItem
 					!= item.html
 	
-	if model.grid.export && model.grid.export.length
+	if model.grid.export && model.grid.export.length && permissions.canExport
 		.export
 			if model.grid.export.length === 1
 				for exp in model.grid.export
-					if exp.enabled
+					if exp.enabled && (exp.permission && permissions[exp.permission] || !exp.permission)
 						- attributes = {}
 						if exp.action === 'export'
 							- attributes['data-target'] = '#exportModal'
@@ -70,7 +73,7 @@
 						span.caret
 					ul.dropdown-menu(role='menu')
 						for exp in model.grid.export
-							if exp.enabled
+							if exp.enabled && (exp.permission && permissions[exp.permission] || !exp.permission)
 								- attributes = {}
 								if exp.action === 'export'
 									- attributes['data-target'] = '#exportModal'
